requirements.txt
python-dotenv
Django
fastapi
pydantic
uvicorn
psycopg2
aiohttp
requests
pillow
python-multipart
